+delete card rights

+++ controllers/cards.js
@@ -1,4 +1,7 @@
 const Card = require('../models/card');
+const ForbiddenError = require('../errors/ForbiddenError');
+const BadRequestError = require('../errors/BadRequestError');
+const NotFoundError = require('../errors/NotFoundError');
 
 const getCards = (req, res) => {
   Card.find({})
@@ -27,19 +30,28 @@ const createCard = (req, res) => {
     });
 };
 
-const deleteCard = (req, res) => {
-  Card.findByIdAndRemove(req.params.cardId)
-    .orFail(() => new Error('Карточка не найдена'))
+const deleteCard = (req, res, next) => {
+  const userId = req.user._id;
+
+  Card.findById(req.params.cardId)
+    .orFail(() => {
+      throw new NotFoundError('Карточка не найдена');
+    })
     .then((data) => {
-      res.status(200).send(data);
+      if (!data.owner.equals(userId)) {
+        throw new ForbiddenError('Нельзя удалить чужую карточку');
+      } else {
+        Card.findByIdAndRemove(req.params.cardId)
+          .then((card) => {
+            res.status(200).send({ card });
+          });
+      }
     })
     .catch((err) => {
       if (err.name === 'CastError') {
-        res.status(400).send({ message: 'Карточка не найдена' });
-      } else if (err.message === 'Карточка не найдена') {
-        res.status(404).send({ message: 'Карточка не найдена' });
+        next(new BadRequestError('Введены некорректные данные'));
       } else {
-        res.status(500).send({ message: 'Произошла ошибка' });
+        next(err);
       }
     });
 };
@@ -87,3 +99,5 @@ module.exports = {
   putLike,
   deleteLike,
 };
+
+// TODO // досоздать оставшиеся ошибки // прописать ошибки всем
